[refactor] types/ 경로로 Mypage에서 사용하는 Match type들 분리
